requirements.txt
streamlit
python-slugify
htbuilder
